Add Authorized for Admin and User

+++ EmployeeMVC/Controllers/EmployeeController.cs
@@ -20,6 +20,14 @@ public class EmployeeController : Controller
     }
     public IActionResult Index()
     {
+        if (HttpContext.Session.GetString("email") == null)
+        {
+            return RedirectToAction("Unauthorized", "Error");
+        }
+        if (HttpContext.Session.GetString("role") != "Admin")
+        {
+            return RedirectToAction("Forbidden", "Error");
+        }
         var employee = employeeRepository.GetAll()
             .Select(e => new EmployeeVM
             {
@@ -38,6 +46,14 @@ public class EmployeeController : Controller
     //Get
     public IActionResult Details(string NIK)
     {
+        if (HttpContext.Session.GetString("email") == null)
+        {
+            return RedirectToAction("Unauthorized", "Error");
+        }
+        if (HttpContext.Session.GetString("role") != "Admin")
+        {
+            return RedirectToAction("Forbidden", "Error");
+        }
         var employee = employeeRepository.GetById(NIK);
         return View(new EmployeeVM
         {
@@ -55,6 +71,14 @@ public class EmployeeController : Controller
     //Get
     public IActionResult Create()
     {
+        if (HttpContext.Session.GetString("email") == null)
+        {
+            return RedirectToAction("Unauthorized", "Error");
+        }
+        if (HttpContext.Session.GetString("role") != "Admin")
+        {
+            return RedirectToAction("Forbidden", "Error");
+        }
         return View();
     }
 
@@ -62,6 +86,14 @@ public class EmployeeController : Controller
     [ValidateAntiForgeryToken]
     public IActionResult Create(EmployeeVM employee)
     {
+        if (HttpContext.Session.GetString("email") == null)
+        {
+            return RedirectToAction("Unauthorized", "Error");
+        }
+        if (HttpContext.Session.GetString("role") != "Admin")
+        {
+            return RedirectToAction("Forbidden", "Error");
+        }
         var result = employeeRepository.Insert(new Employee
         {
             NIK = employee.NIK,
@@ -81,6 +113,14 @@ public class EmployeeController : Controller
     //Get
     public IActionResult Edit(string NIK)
     {
+        if (HttpContext.Session.GetString("email") == null)
+        {
+            return RedirectToAction("Unauthorized", "Error");
+        }
+        if (HttpContext.Session.GetString("role") != "Admin")
+        {
+            return RedirectToAction("Forbidden", "Error");
+        }
         var employee = context.Employees.Find(NIK);
         return View(new EmployeeVM
         {
@@ -99,6 +139,14 @@ public class EmployeeController : Controller
     [ValidateAntiForgeryToken]
     public IActionResult Edit(EmployeeVM employee)
     {
+        if (HttpContext.Session.GetString("email") == null)
+        {
+            return RedirectToAction("Unauthorized", "Error");
+        }
+        if (HttpContext.Session.GetString("role") != "Admin")
+        {
+            return RedirectToAction("Forbidden", "Error");
+        }
         var result = employeeRepository.Update(new Employee
         {
             NIK = employee.NIK,
@@ -118,6 +166,14 @@ public class EmployeeController : Controller
     }
     public IActionResult Delete(string NIK)
     {
+        if (HttpContext.Session.GetString("email") == null)
+        {
+            return RedirectToAction("Unauthorized", "Error");
+        }
+        if (HttpContext.Session.GetString("role") != "Admin")
+        {
+            return RedirectToAction("Forbidden", "Error");
+        }
         var employee = employeeRepository.GetById(NIK);
         return View(new EmployeeVM
         {
@@ -136,6 +192,14 @@ public class EmployeeController : Controller
     [ValidateAntiForgeryToken]
     public IActionResult Remove(string NIK)
     {
+        if (HttpContext.Session.GetString("email") == null)
+        {
+            return RedirectToAction("Unauthorized", "Error");
+        }
+        if (HttpContext.Session.GetString("role") != "Admin")
+        {
+            return RedirectToAction("Forbidden", "Error");
+        }
         var result = employeeRepository.Delete(NIK);
         if (result > 0)
         {
